team_ginkgo/coursework_two/pyproject.toml
[project]
name = "pythonproject1"
version = "0.1.0"
description = ""
authors = [
    {name = "hszli46",email = "hszbli46@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pymupdf (>=1.25.5,<2.0.0)",
    "selenium (>=4.31.0,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "sphinx (>=8.2.3,<9.0.0)",
    "sphinx-rtd-theme (>=3.0.2,<4.0.0)",
    "sphinx-autodoc-typehints (>=3.2.0,<4.0.0)",
    "sphinxcontrib-plantuml (>=0.30,<0.31)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
sphinx = "^8.2.3"

